+++ src/main/java/de/php_perfect/intellij/ddev/cmd/DockerImpl.java
@@ -2,6 +2,7 @@ package de.php_perfect.intellij.ddev.cmd;
 
 import com.intellij.execution.ExecutionException;
 import com.intellij.execution.configurations.GeneralCommandLine;
+import org.jetbrains.annotations.NotNull;
 
 public final class DockerImpl implements Docker {
     @Override
@@ -15,6 +16,20 @@ public final class DockerImpl implements Docker {
             ).getExitCode() == 0;
         } catch (ExecutionException e) {
             return false;
+        }
+    }
+
+    @Override
+    public @NotNull String getContext(String workDirectory) {
+        try {
+            return ProcessExecutor.getInstance().executeCommandLine(
+                    new GeneralCommandLine("docker", "context", "show")
+                            .withWorkDirectory(workDirectory),
+                    5_000,
+                    false
+            ).getStdout();
+        } catch (ExecutionException e) {
+            return "";
         }
     }
 }