+++ src/main/java/de/php_perfect/intellij/ddev/notification/DdevNotifierImpl.java
@@ -123,12 +123,12 @@ public final class DdevNotifierImpl implements DdevNotifier {
     }
 
     @Override
-    public void notifyDockerNotAvailable() {
+    public void notifyDockerNotAvailable(final @NotNull String context) {
         ApplicationManager.getApplication().invokeLater(() -> NotificationGroupManager.getInstance()
                 .getNotificationGroup(STICKY)
                 .createNotification(
                         DdevIntegrationBundle.message("notification.dockerNotAvailable.title"),
-                        DdevIntegrationBundle.message("notification.dockerNotAvailable.text"),
+                        DdevIntegrationBundle.message("notification.dockerNotAvailable.text", context),
                         NotificationType.WARNING
                 )
                 .addAction(new ReloadPluginAction())