+++ src/main/java/de/php_perfect/intellij/ddev/state/DdevStateManagerImpl.java
@@ -41,7 +41,7 @@ public final class DdevStateManagerImpl implements DdevStateManager {
     public void initialize(boolean reinitialize) {
         if (!reinitialize && !Docker.getInstance().isRunning(this.project.getBasePath())) {
             LOG.debug("Docker not available. Skipping initialization");
-            DdevNotifier.getInstance(this.project).notifyDockerNotAvailable();
+            DdevNotifier.getInstance(this.project).notifyDockerNotAvailable(Docker.getInstance().getContext(this.project.getBasePath()));
 
             return;
         }